[hotfix] 모바일 레이아웃 수정
