+++ webapp/src/component/HomeNavigation/index.tsx
@@ -4,6 +4,7 @@ import {
   HomeOutlined,
   CheckSquareOutlined,
   CalendarOutlined,
+  ClockCircleOutlined,
 } from "@ant-design/icons";
 import { Link, useLocation } from "react-router-dom";
 
@@ -12,19 +13,19 @@ export const HomeNavigation = () => {
 
   return (
     <Menu mode="horizontal" selectedKeys={[location.pathname]}>
-      <Menu.Item key="/">
-        <Link to="/">
-          <HomeOutlined /> Home Funny
-        </Link>
-      </Menu.Item>
       <Menu.Item key="/bath">
         <Link to="/bath">
-          <CalendarOutlined /> Bath
+          <CalendarOutlined /> 화장실
         </Link>
       </Menu.Item>
       <Menu.Item key="/wordtest">
         <Link to="/wordtest">
-          <CheckSquareOutlined /> Word Test
+          <CheckSquareOutlined /> 단어시험
+        </Link>
+      </Menu.Item>
+      <Menu.Item key="/timeline">
+        <Link to="/timeline">
+          <ClockCircleOutlined /> 스케줄
         </Link>
       </Menu.Item>
     </Menu>
